Fix critical issue for page_size overflow + tests
A message larger than the maximal page size will block the entire
buffer until it would fill up and drop the entry that's too large.

This patch fixes the issue (and adds corresponding tests) so that
a message too large for the page is left behind, as the app currently
does, except if the page is currently empty, in which case, the message
will be forwarded alone.

This will prompt for a behaviour where a message that is way too large
will always be sent in its own page.